Application Folder/pyproject.toml
[project]
name = "repl-nix-aquarestoreai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask==2.3.2",
    "flask-sqlalchemy==3.0.3",
    "psycopg2-binary>=2.9.9",
    "streamlit>=1.39.0",
    "openai>=1.51.2",
    "google-cloud-speech==2.21.0",
    "google-cloud-vision==3.4.4",
    "en-core-web-sm==3.5.0",
    "thinc==8.1.10",
    "spacy==3.5.0",
    "numpy==1.24.3",
    "opencv-python-headless==4.7.0.72",
    "flask-socketio>=5.4.1",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl" }
